refactor: Align /constant file structure with /component
